+++ src/webviews/_integration/observability/reportObserverError.ts
